Fix stream scanning to not miss boundaries
Occasionally we would see multiple carriage returns and newlines after
the boundary marker. Since the code was using `HasSuffix`, this would
then be false, and the goroutine/Exec call would be stuck forever.

With this approach we just look for the boundary marker, not caring if
it the end of the output buffer. We never determined why this behavior
was happening but it appears to be non-deterministic.

Signed-off-by: Gab Satchi <gsatchithanantham@pivotal.io>